pyproject.toml
[project]
name = "r2r-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "r2r-bo",
    "sqlalchemy>=2.0.43",
]

[tool.uv.sources]
r2r-bo = { path = "../r2r-bo" }


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


